:+1: Use double string-quotes

+++ .stylelintrc.js
@@ -12,6 +12,6 @@ module.exports = {
     },
   ],
   rules: {
-    "string-quotes": ["single"],
+    "string-quotes": ["double"],
   },
 };
